agent/requirements.txt
"""
Agent system requirements
Python packages needed for ML agent system
"""

# Core Agent & LLM
anthropic>=0.7.0                    # Claude API client
langchain>=0.1.0                    # Agent framework (optional, for expansion)
langchain-anthropic>=0.1.0          # LangChain integration

# Database & Data
google-cloud-bigquery>=3.13.0       # BigQuery connector
google-auth-oauthlib>=1.1.0         # GCP authentication
pandas>=2.0.0                       # Data manipulation
pyarrow>=13.0.0                     # Parquet file support

# ML Libraries (leverage existing)
scikit-learn>=1.3.0                 # ML utilities
xgboost>=2.0.0                      # XGBoost models
numpy>=1.24.0                       # Numerical computing

# Vector DB & Embeddings (future enhancement)
chromadb>=0.4.0                     # Vector database (optional MVP)
sentence-transformers>=2.2.0        # Embeddings (optional)

# Baseball Data & APIs
requests>=2.31.0                    # HTTP requests for APIs
beautifulsoup4>=4.12.0              # Web scraping
lxml>=4.9.0                         # XML/HTML parsing

# Local LLM Support (optional)
ollama>=0.1.0                       # Ollama client (optional, for local LLM)

# Utilities
python-dotenv>=1.0.0                # Environment variable loading
pydantic>=2.0.0                     # Data validation
click>=8.1.0                        # CLI framework
streamlit>=1.28.0                   # Dashboard (optional)

# Logging & Monitoring
structlog>=23.1.0                   # Structured logging
prometheus-client>=0.17.0           # Metrics (optional)

# Testing & Development (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
